+++ packages/expo-ui/CHANGELOG.md
@@ -35,6 +35,7 @@ _This version does not introduce any user-facing changes._
 ### 🎉 New features
 
 - [Android] Add compose modifier support. ([#38630](https://github.com/expo/expo/pull/38630) by [@aleqsio](https://github.com/aleqsio))
+- [Android] Add Carousel component for Android. ([#39099](https://github.com/expo/expo/pull/39099) by [@hryhoriiK97](https://github.com/hryhoriiK97))
 
 - [iOS] Add `glassEffect` modifier. ([#38876](https://github.com/expo/expo/pull/38876) by [@nishan](https://github.com/intergalacticspacehighway))